+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -11,6 +11,7 @@ Method call will now require an 'id' field to be present in the document body.
 - Marked the GetAuthorizationHeader method for deprecation since it will no longer be public in a future release.
 - Added samples showing how to configure retry options for both the sync and async clients.
 - Deprecated the `connection_retry_policy` and `retry_options` options in the sync client.
+- Added user warning to non-query methods trying to use `populate_query_metrics` options.
 
 ### 4.3.0b4 (2022-04-07)